python-backend/requirements.txt
fastapi
uvicorn
python-multipart
rembg[cpu]
Pillow
firebase-admin
numpy
opencv-contrib-python-headless
requests
pytz